moved func to common
